+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - change semantics of `dc_get_webxdc_status_updates()` second parameter
   and remove update-id from `DC_EVENT_WEBXDC_STATUS_UPDATE` #3081
 
+### Fixes
+- do not delete messages without Message-IDs as duplicates #3095
+
 ### Changes
 - add more SMTP logging #3093
 - place common headers like `From:` before the large `Autocrypt:` header #3079